server/requirements.txt
fastapi
uvicorn
pyotp
python-dotenv
